Fix delegates not loading because space strategies not available
